backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-admin-interface==0.29.4
django-colorfield==0.12.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pillow==11.1.0
PyJWT==2.10.1
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
text-unidecode==1.3
uritemplate==4.1.1
users==1.0.dev0
